orderList에 사용자 핸드폰 정보 추가

+++ src/main/java/com/spharosacademy/project/SSGBack/order/dto/request/OrdersOptioninputDto.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class OrdersOptioninputDto {
 
     private Long optionListId;
-    private int qty;
+    private long qty;
 
 }
